+++ pages/[province].js
@@ -15,7 +15,9 @@ import HospitalCard from '../components/HospitalCard'
 
 function ProvincePage(props) {
   const { province } = props
-  const { bedFull, hospitalList } = useHospitalDataByProvince(province)
+  const [lat, lon] = (geprops.geo ?? '').split(',')
+  const geo = props.geo ? { lat, lon } : null
+  const { bedFull, hospitalList } = useHospitalDataByProvince(province, geo)
 
   const isLoading = !Boolean(hospitalList)
 
@@ -80,8 +82,8 @@ function ProvincePage(props) {
 }
 
 ProvincePage.getInitialProps = async ({ query }) => {
-  const { province } = query
-  return { province }
+  const { province, geo } = query
+  return { province, geo }
 }
 
 export default ProvincePage
